🥸 Push libft

+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: databey <databey@student.42kocaeli.com.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/09 17:46:11 by databey           #+#    #+#             */
-/*   Updated: 2023/10/09 17:58:12 by databey          ###   ########.fr       */
+/*   Updated: 2023/10/10 11:25:06 by databey          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (dstsize == 0)
